Optimize loginSubmit, and global url
Removed event (e) argument from inline function, as it is unnecessary. Changed global url variable in server file and in url files. Now server contains an or (||) to select whether to run in dev mode or prod mode. When files are uploaded to server, url files will be erased, and server will default to dev mode.