Solve database failure to create table because of multiple recipe matches.
Database refused to create table when recipe was posted multiple times with same blank fields. This was because the sql code would match multiple rows in 'recipes', and then it would return the id of the first match, which was a row that already had an ingredients table created for it. Thus the database would fail to create the table, and return an error that the table already existed. This failure may have been one of the things that restarted the server, and logged the user out.